requirements.txt
numpy
pyaudio
python-dotenv
groq
pydub
asyncio
cartesia